+++ libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/internal/ui/RouteArrowComponentTest.kt
@@ -3,11 +3,11 @@ package com.mapbox.navigation.ui.maps.internal.ui
 import android.content.Context
 import androidx.appcompat.content.res.AppCompatResources
 import com.mapbox.bindgen.ExpectedFactory
-import com.mapbox.maps.MapView
 import com.mapbox.maps.MapboxMap
 import com.mapbox.maps.Style
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
+import com.mapbox.navigation.core.directions.session.RoutesObserver
 import com.mapbox.navigation.core.trip.session.RouteProgressObserver
 import com.mapbox.navigation.testing.MainCoroutineRule
 import com.mapbox.navigation.ui.maps.route.arrow.api.MapboxRouteArrowApi
@@ -16,7 +16,9 @@ import com.mapbox.navigation.ui.maps.route.arrow.model.ClearArrowsValue
 import com.mapbox.navigation.ui.maps.route.arrow.model.InvalidPointError
 import com.mapbox.navigation.ui.maps.route.arrow.model.RouteArrowOptions
 import com.mapbox.navigation.ui.maps.route.arrow.model.UpdateManeuverArrowValue
+import io.mockk.Runs
 import io.mockk.every
+import io.mockk.just
 import io.mockk.mockk
 import io.mockk.mockkStatic
 import io.mockk.slot
@@ -28,10 +30,9 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-@OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
+@OptIn(ExperimentalPreviewMapboxNavigationAPI::class, ExperimentalCoroutinesApi::class)
 class RouteArrowComponentTest {
 
-    @OptIn(ExperimentalCoroutinesApi::class)
     @get:Rule
     var coroutineRule = MainCoroutineRule()
 
@@ -53,62 +54,86 @@ class RouteArrowComponentTest {
     }
 
     @Test
-    fun `route progress test`() {
+    fun `route progress test`() = coroutineRule.runBlockingTest {
         val expected = ExpectedFactory.createError<InvalidPointError, UpdateManeuverArrowValue>(
             InvalidPointError("", null)
         )
         val routeProgressObserverSlot = slot<RouteProgressObserver>()
-        val mockMapboxNavigation = mockk<MapboxNavigation>(relaxed = true)
+        val mockMapboxNavigation = mockk<MapboxNavigation>(relaxed = true) {
+            every { registerRoutesObserver(any()) } just Runs
+            every { registerRouteProgressObserver(capture(routeProgressObserverSlot)) } just Runs
+        }
         val mockApi = mockk<MapboxRouteArrowApi> {
             every { addUpcomingManeuverArrow(any()) } returns expected
         }
         val mockStyle = mockk<Style>()
-        val mockMap = mockk<MapboxMap> {
-            every { getStyle() } returns mockStyle
-        }
-        val mockMapView = mockk<MapView>() {
-            every { getMapboxMap() } returns mockMap
-        }
         val mockView = mockk<MapboxRouteArrowView>(relaxed = true)
-        RouteArrowComponent(
-            mockMapView,
+        val sut = RouteArrowComponent(
+            mockMapWithStyleLoaded(mockStyle),
             routeArrowOptions,
             mockApi,
             mockView
-        ).onAttached(mockMapboxNavigation)
-        verify {
-            mockMapboxNavigation.registerRouteProgressObserver(capture(routeProgressObserverSlot))
-        }
+        )
 
+        sut.onAttached(mockMapboxNavigation)
         routeProgressObserverSlot.captured.onRouteProgressChanged(mockk())
 
         verify { mockView.renderManeuverUpdate(mockStyle, expected) }
     }
 
     @Test
-    fun `onDetached should clear all route arrows`() {
-        val style = mockk<Style>()
-        val mockMapView = mockk<MapView> {
-            every { getMapboxMap() } returns mockk {
-                every { getStyle() } returns style
-            }
+    fun `should clear route arrows when routes are cleared`() = coroutineRule.runBlockingTest {
+        val routesObserverSlot = slot<RoutesObserver>()
+        val mockMapboxNavigation = mockk<MapboxNavigation>(relaxed = true) {
+            every { registerRoutesObserver(capture(routesObserverSlot)) } just Runs
+            every { registerRouteProgressObserver(any()) } just Runs
         }
+        val mockStyle = mockk<Style>()
+        val clearValue = mockk<ClearArrowsValue>()
+        val mockApi = mockk<MapboxRouteArrowApi> {
+            every { clearArrows() } returns clearValue
+        }
+        val mockView = mockk<MapboxRouteArrowView>(relaxed = true)
+        val sut = RouteArrowComponent(
+            mockMapWithStyleLoaded(mockStyle),
+            routeArrowOptions,
+            mockApi,
+            mockView
+        )
+
+        sut.onAttached(mockMapboxNavigation)
+        routesObserverSlot.captured.onRoutesChanged(
+            mockk {
+                every { navigationRoutes } returns emptyList()
+            }
+        )
+
+        verify { mockView.render(mockStyle, clearValue) }
+    }
+
+    @Test
+    fun `onDetached should clear all route arrows`() = coroutineRule.runBlockingTest {
+        val mockStyle = mockk<Style>()
         val clearValue = mockk<ClearArrowsValue>()
         val mockApi = mockk<MapboxRouteArrowApi> {
             every { clearArrows() } returns clearValue
         }
         val mockView = mockk<MapboxRouteArrowView>(relaxed = true)
         val mockMapboxNavigation = mockk<MapboxNavigation>(relaxed = true)
-
         val sut = RouteArrowComponent(
-            mockMapView,
+            mockMapWithStyleLoaded(mockStyle),
             routeArrowOptions,
             mockApi,
             mockView
         )
+
         sut.onAttached(mockMapboxNavigation)
         sut.onDetached(mockMapboxNavigation)
 
-        verify { mockView.render(style, clearValue) }
+        verify { mockView.render(mockStyle, clearValue) }
+    }
+
+    private fun mockMapWithStyleLoaded(style: Style): MapboxMap = mockk {
+        every { getStyle() } returns style
     }
 }